Title: 최솟값 3개, Time: 1064ms, Memory: 29MB, Status: Passed - Codetree

+++ 250513/최솟값 3개/top-3-smallest-number.java
@@ -38,7 +38,7 @@ class Solver {
             int min1 = priorityQueue.poll();
             int min2 = priorityQueue.poll();
             int min3 = priorityQueue.poll();
-            System.out.println(min1 * min2 * min3);
+            System.out.println((long) min1 * min2 * min3);
             priorityQueue.add(min1);
             priorityQueue.add(min2);
             priorityQueue.add(min3);
